requirements.txt
psycopg2-binary
pandas
odfpy
requests